+++ client/src/main/java/net/luxvacuos/voxel/client/rendering/api/opengl/objects/PointLight.java
@@ -20,11 +20,11 @@
 
 package net.luxvacuos.voxel.client.rendering.api.opengl.objects;
 
-import net.luxvacuos.igl.vector.Vector3f;
+import net.luxvacuos.igl.vector.Vector3d;
 
 public class PointLight extends Light {
 
-	public PointLight(Vector3f position) {
+	public PointLight(Vector3d position) {
 		super(position);
 	}
